pyproject.toml
[tool.poetry]
name = "dataneuron"
version = "0.3.1"
description = "An AI Data framework to create AI Data Analyst"
authors = ["Vysakh Sreenivasan <srvysakh@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dataneuron", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
pyyaml = "^6.0"
openai = "^1.35.15"
pyaml = "^24.7.0"
colorama = "^0.4.6"
psycopg2-binary = { version = "^2.9.0", optional = true }
mysql-connector-python = { version = "^8.0.0", optional = true }
pyodbc = { version = "^4.0.0", optional = true }
sqlparse = "^0.5.1"
tabulate = "^0.9.0"
requests = "^2.32.3"
duckdb = { version = "^0.9.0", optional = true }
pdfkit = { version = "^1.0.0" , optional = true }
clickhouse-connect = { version ="^0.7.17",   optional = true }
flask = "^3.0.3"

[tool.poetry.extras]
sqlite = [] 
postgres = ["psycopg2-binary"]
mysql = ["mysql-connector-python"]
mssql = ["pyodbc"]
csv = ["duckdb"]
clickhouse = ["clickhouse-connect"]  
pdf = ["pdfkit"]
all = ["psycopg2-binary", "mysql-connector-python", "pyodbc", "duckdb", "clickhouse-connect", "pdfkit"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dnn = "dataneuron.cli:cli"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
python_files = "test_*.py"